Cambios 18/10/22

+++ database/models/justificacion.model.js
@@ -41,7 +41,7 @@ const JustificacionTableSGASchema = {
     unidadesJustificadas: {
         field: 'unidades_justificadas',
         allowNull: false,
-        type: DataTypes.INTEGER
+        type: DataTypes.FLOAT(11)
     },
 
     createdAt: {
